[Chore] #245 여행 후기 평점 type 변경

+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/domain/model/ScheduleReviewInfo.kt
@@ -3,6 +3,6 @@ package kr.tekit.lion.daongil.domain.model
 data class ScheduleReviewInfo(
     val reviewId: Long,
     val content: String,
-    val grade: Double,
+    val grade: Float,
     val imageList: List<String>?,
 )
